% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_loss_history}
\alias{compute_loss_history}
\title{Compute Training Loss History}
\usage{
compute_loss_history(boost_obj)
}
\arguments{
\item{boost_obj}{A boosted model object from boost_regression()}
}
\value{
Vector of MSE values (length B+1), starting with initial MSE
}
\description{
Calculate MSE on training data at each boosting iteration.
}
