% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_boost}
\alias{predict_boost}
\title{Predict using a Boosted Model}
\usage{
predict_boost(boost_obj, X_new)
}
\arguments{
\item{boost_obj}{A boosted model object from boost_regression()}

\item{X_new}{Feature matrix for prediction (m x p)}
}
\value{
Vector of predictions (length m)
}
\description{
Generate predictions from a boosted model on new data.
}
