\name{unifiedml-package}
\alias{unifiedml-package}
\alias{unifiedml}
\docType{package}
\title{\packageTitle{unifiedml}}
\description{\packageDescription{unifiedml}}
\section{Package Content}{\packageIndices{unifiedml}}
\author{\packageAuthor{unifiedml}}
\section{Maintainer}{\packageMaintainer{unifiedml}}
\keyword{package}
